+++ src/animation/utils/transitions.ts
@@ -12,6 +12,7 @@ import { getDefaultTransition } from "./default-transitions"
 import { warning } from "hey-listen"
 import { getAnimatableNone } from "../../render/dom/value-types/animatable-none"
 import { instantAnimationState } from "../../utils/use-instant-transition-state"
+import { resolveFinalValueInKeyframes } from "../../utils/resolve-value"
 
 type StopAnimation = { stop: () => void }
 
@@ -214,9 +215,10 @@ function getAnimation(
     }
 
     function set(): StopAnimation {
-        value.set(target)
+        const finalTarget = resolveFinalValueInKeyframes(target)
+        value.set(finalTarget)
         onComplete()
-        valueTransition?.onUpdate?.(target)
+        valueTransition?.onUpdate?.(finalTarget)
         valueTransition?.onComplete?.()
         return { stop: () => {} }
     }