+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Framer Motion adheres to [Semantic Versioning](http://semver.org/).
 -   Removing `layoutDependency` from forwarded props. [Issue](https://github.com/framer/motion/issues/1350)
 -   `Reorder.Item` correctly fires `onDrag`. [Issue](https://github.com/framer/motion/issues/1348)
 -   Fires `onPressStart` and `onHoverStart` **after** triggering animations.
+-   Replay keyframes when variant changes. [Issue](https://github.com/framer/motion/issues/1346)
+-   Correctly SSR final keyframe when `initial` is `false`.
 
 ## [5.3.0] 2021-11-11